requirements.txt
asgiref==3.3.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
lxml==4.8.0
mysqlclient==2.0.3
Pillow==8.1.0
psycopg2==2.9.1
python-dateutil==1.5
pytz==2021.1
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.1
urllib3==1.26.8
whitenoise==5.2.0
